+++ controllers/routers/admin.js
@@ -1,4 +1,4 @@
-const { _M_ } = require('../middleware/');
+const { _M_ } = require('../modules/');
 const { _U_ } = require('../utilities/');
 const { configsModel, mediaModel, postsModel } = require('../../models/');