+++ services/security.js
@@ -14,13 +14,13 @@ class ContentSecurityPolicy {
 
   rules() {
     const RuleMaps  = {
-      "default-src" : "'self' ",
-      "script-src"  : `'self' 'unsafe-inline' ${this.getTrustedDomains('script')}`,
-      "style-src"   : `'self' 'unsafe-inline' ${this.getTrustedDomains('style')}`,
-      "font-src"    : `${this.getTrustedDomains('font')}`,
-      "img-src"     : "*",
-      "frame-src"   : "'none'",
-      "object-src"  : "'none'",
+      'default-src' : `'self'`,
+      'script-src'  : `'self' 'unsafe-inline' ${this.getTrustedDomains('script')}`,
+      'style-src'   : `'self' 'unsafe-inline' ${this.getTrustedDomains('style')}`,
+      'font-src'    : `${this.getTrustedDomains('font')}`,
+      'img-src'     : "*",
+      'frame-src'   : `'none'`,
+      'object-src'  : `'none'`,
     };
     return Object.entries(RuleMaps).map(i => i.join(' ')).join('; ');
   }
@@ -34,12 +34,12 @@ const securityHeaderAgent = (app) => {
   CSPConfigs.addToWhitelist('script', 'https://code.jquery.com/');
 
   app.set('x-powered-by', false);
-  app.use((req, res, next) => {
+  app.use(function CSPHTTPHeader(req, res, next) {
     res.set({
-      "Content-Security-Policy"  : CSPConfigs.rules(),
-      "Cache-Control"            : "max-age=0",
-      "X-XSS-Protection"         : "1; mode=block",
-      "X-Frame-Options"          : "DENY",
+      'Content-Security-Policy'  : CSPConfigs.rules(),
+      'Cache-Control'            : 'max-age=0',
+      'X-XSS-Protection'         : '1; mode=block',
+      'X-Frame-Options'          : 'DENY',
     });
     return next();
   });